Refactor THE-846 object lifecycle manager
