+++ api-go/internal/manager/s3_object.go
@@ -2,19 +2,12 @@ package manager
 
 import (
 	"context"
-	"crypto/md5"
-	"crypto/sha256"
-	"encoding/hex"
 	"errors"
 	"fmt"
 	"io"
-	"strconv"
-	"strings"
-	"time"
 
 	"github.com/example/sfree/api-go/internal/repository"
 	"go.mongodb.org/mongo-driver/bson/primitive"
-	"go.mongodb.org/mongo-driver/mongo"
 )
 
 var (
@@ -245,523 +238,4 @@ func (s *MultipartCompletionService) CompleteMultipartUpload(ctx context.Context
 
 func (s *MultipartCompletionService) CompleteMultipartUploadRecord(ctx context.Context, bucketID primitive.ObjectID, mu *repository.MultipartUpload, requestedParts []CompleteMultipartPart) (CompleteMultipartUploadResult, error) {
 	return s.service.CompleteMultipartUploadRecord(ctx, bucketID, mu, requestedParts)
-}
-
-func (s *objectService) PutObject(ctx context.Context, bucket *repository.Bucket, name string, body io.Reader, chunkSize int, contentType string, userMetadata map[string]string) (PutObjectResult, error) {
-	sources, err := s.sources.ListByIDs(ctx, bucket.SourceIDs)
-	if err != nil {
-		return PutObjectResult{}, err
-	}
-	if len(bucket.SourceIDs) == 0 {
-		return PutObjectResult{}, ErrNoSources
-	}
-	if err := requireResolvedSources(bucket.SourceIDs, sources); err != nil {
-		return PutObjectResult{}, err
-	}
-
-	chunks, err := s.uploadChunks(ctx, body, sources, chunkSize, SelectorForBucket(bucket, sources))
-	if err != nil {
-		return PutObjectResult{}, err
-	}
-
-	fileDoc := repository.File{
-		BucketID:     bucket.ID,
-		Name:         name,
-		CreatedAt:    s.now(),
-		Chunks:       chunks,
-		ContentType:  contentType,
-		UserMetadata: cloneStringMap(userMetadata),
-	}
-	fileDoc.ETag = newObjectETag(fileDoc)
-	currentFile, previousFile, err := s.files.ReplaceByName(ctx, fileDoc)
-	if err != nil {
-		_ = s.deleteChunks(ctx, chunks)
-		return PutObjectResult{}, err
-	}
-	var cleanupErr error
-	if previousFile != nil {
-		cleanupErr = s.deleteFileChunksIfUnreferenced(ctx, previousFile.Chunks)
-	}
-	return PutObjectResult{File: *currentFile, CleanupErr: cleanupErr}, nil
-}
-
-func (s *objectService) UploadMultipartPartRecord(ctx context.Context, bucket *repository.Bucket, mu *repository.MultipartUpload, partNumber int, body io.Reader, chunkSize int) (UploadMultipartPartResult, error) {
-	if bucket == nil || mu == nil || mu.BucketID != bucket.ID {
-		return UploadMultipartPartResult{}, ErrMultipartUploadNotFound
-	}
-
-	sources, err := s.sources.ListByIDs(ctx, bucket.SourceIDs)
-	if err != nil {
-		return UploadMultipartPartResult{}, err
-	}
-	if len(bucket.SourceIDs) == 0 {
-		return UploadMultipartPartResult{}, ErrNoSources
-	}
-	if err := requireResolvedSources(bucket.SourceIDs, sources); err != nil {
-		return UploadMultipartPartResult{}, err
-	}
-
-	chunks, err := s.uploadChunks(ctx, body, sources, chunkSize, SelectorForBucket(bucket, sources))
-	if err != nil {
-		return UploadMultipartPartResult{}, err
-	}
-
-	var totalSize int64
-	for _, ch := range chunks {
-		totalSize += ch.Size
-	}
-	part := repository.UploadPart{
-		PartNumber: partNumber,
-		ETag:       multipartPartETag(chunks),
-		Size:       totalSize,
-		Chunks:     chunks,
-	}
-
-	previous, err := s.multipart.SetPart(ctx, mu.UploadID, part)
-	if err != nil {
-		_ = s.deleteChunks(ctx, chunks)
-		if err == mongo.ErrNoDocuments {
-			return UploadMultipartPartResult{}, ErrMultipartUploadNotFound
-		}
-		return UploadMultipartPartResult{}, err
-	}
-
-	var cleanupErr error
-	if previous != nil {
-		cleanupErr = s.deleteChunks(ctx, previous.Chunks)
-	}
-	return UploadMultipartPartResult{Part: part, ETag: part.ETag, CleanupErr: cleanupErr}, nil
-}
-
-func requireResolvedSources(ids []primitive.ObjectID, sources []repository.Source) error {
-	resolved := make(map[primitive.ObjectID]struct{}, len(sources))
-	for _, source := range sources {
-		resolved[source.ID] = struct{}{}
-	}
-	missing := make([]primitive.ObjectID, 0)
-	for _, id := range ids {
-		if _, ok := resolved[id]; !ok {
-			missing = append(missing, id)
-		}
-	}
-	if len(missing) > 0 {
-		return repository.SourcesNotFoundError{IDs: missing}
-	}
-	return nil
-}
-
-func (s *objectService) CopyObject(ctx context.Context, sourceBucket, destBucket *repository.Bucket, sourceKey, destKey string) (CopyObjectResult, error) {
-	if sourceBucket.ID != destBucket.ID && sourceBucket.UserID != destBucket.UserID {
-		return CopyObjectResult{}, ErrAccessDenied
-	}
-
-	sourceFile, err := s.files.GetByName(ctx, sourceBucket.ID, sourceKey)
-	if err != nil {
-		if err == mongo.ErrNoDocuments {
-			return CopyObjectResult{}, ErrObjectNotFound
-		}
-		return CopyObjectResult{}, err
-	}
-
-	chunks := append([]repository.FileChunk(nil), sourceFile.Chunks...)
-	copyFile := repository.File{
-		BucketID:     destBucket.ID,
-		Name:         destKey,
-		CreatedAt:    s.now(),
-		Chunks:       chunks,
-		ContentType:  sourceFile.ContentType,
-		UserMetadata: cloneStringMap(sourceFile.UserMetadata),
-	}
-	copyFile.ETag = copyObjectETag(*sourceFile)
-	currentFile, previousFile, err := s.files.ReplaceByName(ctx, copyFile)
-	if err != nil {
-		return CopyObjectResult{}, err
-	}
-	var cleanupErr error
-	if previousFile != nil {
-		cleanupErr = s.deleteFileChunksIfUnreferenced(ctx, previousFile.Chunks)
-	}
-	return CopyObjectResult{File: *currentFile, CleanupErr: cleanupErr}, nil
-}
-
-func (s *objectService) DeleteObject(ctx context.Context, bucketID primitive.ObjectID, name string) (DeleteObjectResult, error) {
-	fileDoc, err := s.files.GetByName(ctx, bucketID, name)
-	if err != nil {
-		if err == mongo.ErrNoDocuments {
-			return DeleteObjectResult{}, nil
-		}
-		return DeleteObjectResult{}, err
-	}
-	if err := s.files.Delete(ctx, fileDoc.ID); err != nil {
-		if err == mongo.ErrNoDocuments {
-			return DeleteObjectResult{}, nil
-		}
-		return DeleteObjectResult{}, err
-	}
-	return DeleteObjectResult{
-		Deleted:    true,
-		CleanupErr: s.deleteFileChunksIfUnreferenced(ctx, fileDoc.Chunks),
-	}, nil
-}
-
-func (s *objectService) DeleteFile(ctx context.Context, bucketID, fileID primitive.ObjectID) (DeleteObjectResult, error) {
-	fileDoc, err := s.files.GetByID(ctx, fileID)
-	if err != nil {
-		if err == mongo.ErrNoDocuments {
-			return DeleteObjectResult{}, ErrObjectNotFound
-		}
-		return DeleteObjectResult{}, err
-	}
-	if fileDoc.BucketID != bucketID {
-		return DeleteObjectResult{}, ErrObjectNotFound
-	}
-	if err := s.files.Delete(ctx, fileDoc.ID); err != nil {
-		if err == mongo.ErrNoDocuments {
-			return DeleteObjectResult{}, ErrObjectNotFound
-		}
-		return DeleteObjectResult{}, err
-	}
-	return DeleteObjectResult{
-		Deleted:    true,
-		CleanupErr: s.deleteFileChunksIfUnreferenced(ctx, fileDoc.Chunks),
-	}, nil
-}
-
-func (s *objectService) DeleteBucketContents(ctx context.Context, bucketID primitive.ObjectID) (DeleteBucketContentsResult, error) {
-	files, err := s.files.ListByBucket(ctx, bucketID)
-	if err != nil {
-		return DeleteBucketContentsResult{}, err
-	}
-
-	var uploads []repository.MultipartUpload
-	if s.multipart != nil {
-		uploads, err = s.multipart.ListByBucket(ctx, bucketID)
-		if err != nil {
-			return DeleteBucketContentsResult{}, err
-		}
-	}
-
-	chunks := bucketCleanupChunks(files, uploads)
-	if err := s.files.DeleteByBucket(ctx, bucketID); err != nil {
-		return DeleteBucketContentsResult{}, err
-	}
-	if s.multipart != nil {
-		if err := s.multipart.DeleteByBucket(ctx, bucketID); err != nil {
-			return DeleteBucketContentsResult{}, err
-		}
-	}
-	if err := s.deleteBucketChunksIfUnreferenced(ctx, bucketID, chunks); err != nil {
-		return DeleteBucketContentsResult{}, err
-	}
-	return DeleteBucketContentsResult{
-		FilesDeleted:            len(files),
-		MultipartUploadsDeleted: len(uploads),
-	}, nil
-}
-
-func (s *objectService) CompleteMultipartUpload(ctx context.Context, bucketID primitive.ObjectID, uploadID string, requestedParts []CompleteMultipartPart) (CompleteMultipartUploadResult, error) {
-	mu, err := s.multipart.GetByUploadID(ctx, uploadID)
-	if err != nil {
-		if err == mongo.ErrNoDocuments {
-			return CompleteMultipartUploadResult{}, ErrMultipartUploadNotFound
-		}
-		return CompleteMultipartUploadResult{}, err
-	}
-	return s.CompleteMultipartUploadRecord(ctx, bucketID, mu, requestedParts)
-}
-
-func (s *objectService) CompleteMultipartUploadRecord(ctx context.Context, bucketID primitive.ObjectID, mu *repository.MultipartUpload, requestedParts []CompleteMultipartPart) (CompleteMultipartUploadResult, error) {
-	if mu == nil {
-		return CompleteMultipartUploadResult{}, ErrMultipartUploadNotFound
-	}
-	if mu.BucketID != bucketID {
-		return CompleteMultipartUploadResult{}, ErrMultipartUploadNotFound
-	}
-	if len(requestedParts) == 0 {
-		return CompleteMultipartUploadResult{}, ErrMultipartUploadHasNoParts
-	}
-	for i := 1; i < len(requestedParts); i++ {
-		if requestedParts[i].PartNumber <= requestedParts[i-1].PartNumber {
-			return CompleteMultipartUploadResult{}, ErrMultipartUploadPartOrder
-		}
-	}
-
-	partMap := make(map[int]repository.UploadPart, len(mu.Parts))
-	for _, p := range mu.Parts {
-		partMap[p.PartNumber] = p
-	}
-
-	var allChunks []repository.FileChunk
-	chunkOrder := 0
-	for _, rp := range requestedParts {
-		up, exists := partMap[rp.PartNumber]
-		if !exists {
-			return CompleteMultipartUploadResult{}, InvalidMultipartPartError{PartNumber: rp.PartNumber, Reason: "not uploaded"}
-		}
-		if strings.Trim(rp.ETag, "\"") != strings.Trim(up.ETag, "\"") {
-			return CompleteMultipartUploadResult{}, InvalidMultipartPartError{PartNumber: rp.PartNumber, Reason: "etag mismatch"}
-		}
-		for _, ch := range up.Chunks {
-			allChunks = append(allChunks, repository.FileChunk{
-				SourceID: ch.SourceID,
-				Name:     ch.Name,
-				Order:    chunkOrder,
-				Size:     ch.Size,
-				Checksum: ch.Checksum,
-			})
-			chunkOrder++
-		}
-	}
-
-	etag := multipartETag(requestedParts, partMap)
-	fileDoc := repository.File{
-		BucketID:     bucketID,
-		Name:         mu.ObjectKey,
-		CreatedAt:    s.now(),
-		Chunks:       allChunks,
-		ETag:         etag,
-		ContentType:  mu.ContentType,
-		UserMetadata: cloneStringMap(mu.UserMetadata),
-	}
-	saved, previousFile, err := s.files.ReplaceByName(ctx, fileDoc)
-	if err != nil {
-		return CompleteMultipartUploadResult{}, err
-	}
-
-	var cleanupErrs []error
-	if previousFile != nil {
-		if err := s.deleteFileChunksIfUnreferenced(ctx, previousFile.Chunks); err != nil {
-			cleanupErrs = append(cleanupErrs, err)
-		}
-	}
-
-	requested := make(map[int]bool, len(requestedParts))
-	for _, rp := range requestedParts {
-		requested[rp.PartNumber] = true
-	}
-	for _, p := range mu.Parts {
-		if !requested[p.PartNumber] {
-			if err := s.deleteChunks(ctx, p.Chunks); err != nil {
-				cleanupErrs = append(cleanupErrs, err)
-			}
-		}
-	}
-	if err := s.multipart.Delete(ctx, mu.UploadID); err != nil {
-		cleanupErrs = append(cleanupErrs, err)
-	}
-
-	return CompleteMultipartUploadResult{
-		File:        *saved,
-		Upload:      *mu,
-		ETag:        etag,
-		CleanupErrs: cleanupErrs,
-	}, nil
-}
-
-func (s *objectService) AbortMultipartUpload(ctx context.Context, uploadID string) error {
-	return AbortMultipartUpload(ctx, s.multipart, s.deleteChunks, uploadID)
-}
-
-func (s *objectService) AbortMultipartUploadRecord(ctx context.Context, mu *repository.MultipartUpload) error {
-	return AbortMultipartUploadRecord(ctx, s.multipart, s.deleteChunks, mu)
-}
-
-func AbortMultipartUpload(ctx context.Context, multipart MultipartUploadAbortStore, deleteChunks func(context.Context, []repository.FileChunk) error, uploadID string) error {
-	mu, err := multipart.GetByUploadID(ctx, uploadID)
-	if err != nil {
-		if err == mongo.ErrNoDocuments {
-			return ErrMultipartUploadNotFound
-		}
-		return err
-	}
-	return AbortMultipartUploadRecord(ctx, multipart, deleteChunks, mu)
-}
-
-func AbortMultipartUploadRecord(ctx context.Context, multipart MultipartUploadAbortStore, deleteChunks func(context.Context, []repository.FileChunk) error, mu *repository.MultipartUpload) error {
-	if mu == nil {
-		return ErrMultipartUploadNotFound
-	}
-
-	var cleanupErrs []error
-	for _, part := range mu.Parts {
-		if err := deleteChunks(ctx, part.Chunks); err != nil {
-			cleanupErrs = append(cleanupErrs, err)
-		}
-	}
-	if err := errors.Join(cleanupErrs...); err != nil {
-		return err
-	}
-	if err := multipart.Delete(ctx, mu.UploadID); err != nil {
-		if err == mongo.ErrNoDocuments {
-			return ErrMultipartUploadNotFound
-		}
-		return err
-	}
-	return nil
-}
-
-func (s *objectService) deleteFileChunksIfUnreferenced(ctx context.Context, chunks []repository.FileChunk) error {
-	return deleteFileChunksIfUnreferenced(ctx, s.files, s.deleteChunks, chunks)
-}
-
-func (s *objectService) deleteBucketChunksIfUnreferenced(ctx context.Context, bucketID primitive.ObjectID, chunks []repository.FileChunk) error {
-	return deleteBucketChunksIfUnreferenced(ctx, bucketID, s.files, s.multipart, s.deleteChunks, chunks)
-}
-
-func DeleteFileChunksIfUnreferenced(ctx context.Context, sourceRepo *repository.SourceRepository, fileStore ChunkReferenceCounter, chunks []repository.FileChunk) error {
-	return DeleteFileChunksIfUnreferencedWithFactory(ctx, sourceRepo, fileStore, chunks, nil)
-}
-
-func DeleteFileChunksIfUnreferencedWithFactory(ctx context.Context, sourceRepo *repository.SourceRepository, fileStore ChunkReferenceCounter, chunks []repository.FileChunk, factory SourceClientFactory) error {
-	return deleteFileChunksIfUnreferenced(ctx, fileStore, func(ctx context.Context, chunks []repository.FileChunk) error {
-		return DeleteFileChunksWithFactory(ctx, sourceRepo, chunks, factory)
-	}, chunks)
-}
-
-func deleteFileChunksIfUnreferenced(ctx context.Context, files ChunkReferenceCounter, deleteChunks objectChunkDeleter, chunks []repository.FileChunk) error {
-	seen := make(map[string]struct{}, len(chunks))
-	for _, chunk := range chunks {
-		ref := chunk.SourceID.Hex() + "/" + chunk.Name
-		if _, ok := seen[ref]; ok {
-			continue
-		}
-		seen[ref] = struct{}{}
-		count, err := files.CountByChunk(ctx, chunk.SourceID, chunk.Name)
-		if err != nil {
-			return err
-		}
-		if count > 0 {
-			continue
-		}
-		if err := deleteChunks(ctx, []repository.FileChunk{chunk}); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func deleteBucketChunksIfUnreferenced(ctx context.Context, bucketID primitive.ObjectID, files BucketScopedChunkReferenceCounter, multipart objectMultipartStore, deleteChunks objectChunkDeleter, chunks []repository.FileChunk) error {
-	seen := make(map[string]struct{}, len(chunks))
-	for _, chunk := range chunks {
-		ref := chunk.SourceID.Hex() + "/" + chunk.Name
-		if _, ok := seen[ref]; ok {
-			continue
-		}
-		seen[ref] = struct{}{}
-		count, err := files.CountByChunkExcludingBucket(ctx, bucketID, chunk.SourceID, chunk.Name)
-		if err != nil {
-			return err
-		}
-		if count > 0 {
-			continue
-		}
-		if multipart != nil {
-			count, err = multipart.CountByPartChunkExcludingBucket(ctx, bucketID, chunk.SourceID, chunk.Name)
-			if err != nil {
-				return err
-			}
-			if count > 0 {
-				continue
-			}
-		}
-		if err := deleteChunks(ctx, []repository.FileChunk{chunk}); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func bucketCleanupChunks(files []repository.File, uploads []repository.MultipartUpload) []repository.FileChunk {
-	var chunks []repository.FileChunk
-	for _, file := range files {
-		chunks = append(chunks, file.Chunks...)
-	}
-	for _, upload := range uploads {
-		for _, part := range upload.Parts {
-			chunks = append(chunks, part.Chunks...)
-		}
-	}
-	return chunks
-}
-
-func ObjectETag(file repository.File) string {
-	if file.ETag != "" {
-		return file.ETag
-	}
-	return legacyObjectETag(file)
-}
-
-func newObjectETag(file repository.File) string {
-	if etag, ok := checksumObjectETag(file.Chunks); ok {
-		return etag
-	}
-	return legacyObjectETag(file)
-}
-
-func copyObjectETag(source repository.File) string {
-	if source.ETag != "" {
-		return source.ETag
-	}
-	if etag, ok := checksumObjectETag(source.Chunks); ok {
-		return etag
-	}
-	return legacyObjectETag(source)
-}
-
-func checksumObjectETag(chunks []repository.FileChunk) (string, bool) {
-	h := sha256.New()
-	for _, chunk := range chunks {
-		if chunk.Checksum == "" {
-			return "", false
-		}
-		_, _ = h.Write([]byte(chunk.Checksum))
-		_, _ = h.Write([]byte(":"))
-		_, _ = h.Write([]byte(strconv.FormatInt(chunk.Size, 10)))
-		_, _ = h.Write([]byte(";"))
-	}
-	return "\"" + hex.EncodeToString(h.Sum(nil)) + "\"", true
-}
-
-func legacyObjectETag(file repository.File) string {
-	h := sha256.New()
-	_, _ = h.Write([]byte(file.Name))
-	_, _ = h.Write([]byte(file.CreatedAt.UTC().Format(time.RFC3339Nano)))
-	for _, chunk := range file.Chunks {
-		_, _ = h.Write([]byte(chunk.SourceID.Hex()))
-		_, _ = h.Write([]byte(chunk.Name))
-		_, _ = h.Write([]byte(strconv.Itoa(chunk.Order)))
-		_, _ = h.Write([]byte(":"))
-		_, _ = h.Write([]byte(strconv.FormatInt(chunk.Size, 10)))
-	}
-	return "\"" + hex.EncodeToString(h.Sum(nil)) + "\""
-}
-
-func multipartPartETag(chunks []repository.FileChunk) string {
-	h := md5.New()
-	for _, chunk := range chunks {
-		_, _ = h.Write([]byte(chunk.Name))
-	}
-	return fmt.Sprintf("\"%s\"", hex.EncodeToString(h.Sum(nil)))
-}
-
-func cloneStringMap(values map[string]string) map[string]string {
-	if len(values) == 0 {
-		return nil
-	}
-	clone := make(map[string]string, len(values))
-	for key, value := range values {
-		clone[key] = value
-	}
-	return clone
-}
-
-func multipartETag(requestedParts []CompleteMultipartPart, partMap map[int]repository.UploadPart) string {
-	h := md5.New()
-	for _, rp := range requestedParts {
-		up := partMap[rp.PartNumber]
-		partHash := strings.Trim(up.ETag, "\"")
-		raw, _ := hex.DecodeString(partHash)
-		_, _ = h.Write(raw)
-	}
-	return fmt.Sprintf("\"%s-%d\"", hex.EncodeToString(h.Sum(nil)), len(requestedParts))
 }
